+++ src/packages/piral-cli/src/apps/common/npm.ts
@@ -1,6 +1,9 @@
 import { exec } from 'child_process';
 import { resolve } from 'path';
+import { createReadStream } from 'fs';
 import { isWindows } from './info';
+import { inspectPackage } from './inspect';
+import { readJson } from './io';
 
 const npmCommand = isWindows ? 'npm.cmd' : 'npm';
 
@@ -18,13 +21,18 @@ export function installDependencies(target = '.', ...flags: Array<string>) {
   return runNpmProcess(['install', ...flags], target);
 }
 
-export function installPackage(name: string, version = 'latest', target = '.', ...flags: Array<string>) {
-  return runNpmProcess(['install', `${name}@${version}`, ...flags], target);
+export function installPackage(packageRef: string, target = '.', ...flags: Array<string>) {
+  return runNpmProcess(['install', packageRef, ...flags], target);
 }
 
 export function createPackage(target = '.', ...flags: Array<string>) {
   return runNpmProcess(['pack', ...flags], target);
 }
+
+export type PackageType = 'registry' | 'file' | 'git';
+
+const gitPrefix = 'git+';
+const filePrefix = 'file:';
 
 /**
  * Looks at the provided package name and normalizes it
@@ -36,12 +44,56 @@ export function createPackage(target = '.', ...flags: Array<string>) {
  * ]
  * @param fullName The provided package name.
  */
-export function dissectPackageName(fullName: string): [string, string, boolean] {
-  const index = fullName.indexOf('@', 1);
+export function dissectPackageName(fullName: string): [string, string, boolean, PackageType] {
+  const localFile = /^[\.\/\~]/.test(fullName);
+  const git = fullName.startsWith(gitPrefix) || /^(https?|ssh):\/\/.*\.git$/.test(fullName);
+  const type = localFile ? 'file' : git ? 'git' : 'registry';
 
-  if (index !== -1) {
-    return [fullName.substr(0, index), fullName.substr(index + 1), true];
+  if (git) {
+    const gitUrl = fullName.startsWith(gitPrefix) ? fullName : `${gitPrefix}${fullName}`;
+    return [gitUrl, 'latest', false, type];
+  } else if (!localFile) {
+    const index = fullName.indexOf('@', 1);
+
+    if (index !== -1) {
+      return [fullName.substr(0, index), fullName.substr(index + 1), true, type];
+    }
   }
 
-  return [fullName, 'latest', false];
+  return [fullName, 'latest', false, type];
+}
+
+export function combinePackageRef(name: string, version: string, type: PackageType) {
+  if (type === 'registry') {
+    return `${name}@${version}`;
+  }
+
+  return name;
+}
+
+export async function getPackageName(root: string, name: string, type: PackageType) {
+  switch (type) {
+    case 'file':
+      const p = resolve(process.cwd(), name);
+      const s = createReadStream(p);
+      const i = await inspectPackage(s);
+      return i.name;
+    case 'git':
+      const pj = await readJson(root, 'package.json');
+      const dd = pj.devDependencies || {};
+      return Object.keys(dd).filter(dep => dd[dep] === name)[0];
+    case 'registry':
+      return name;
+  }
+}
+
+export function getPackageVersion(hadVersion: boolean, sourceName: string, sourceVersion: string, type: PackageType) {
+  switch (type) {
+    case 'registry':
+      return hadVersion && sourceVersion;
+    case 'file':
+      return `file:${sourceName}`;
+    case 'git':
+      return `${sourceName}`;
+  }
 }