+++ src/pages/docs/src/scripts/components/urls.ts
@@ -1,9 +1,18 @@
-const imageBaseUrl = 'https://raw.githubusercontent.com/smapiot/piral/master/docs';
-const docBaseUrl = 'https://github.com/smapiot/piral/tree/master/docs';
+const branch = 'documentation';
+const repo = 'smapiot/piral';
+const docsFolder = 'docs';
+const imageBaseUrl = `https://raw.githubusercontent.com/${repo}/${branch}/${docsFolder}`;
+const docBaseUrl = `https://github.com/${repo}/tree/${branch}/${docsFolder}`;
 
 export function imgRef(path: string) {
   const relPath = path.replace(/^..\//g, '');
-  return `${imageBaseUrl}/${relPath}?sanitize=true`;
+  const url = `${imageBaseUrl}/${relPath}`;
+
+  if (/\.svg$/.test(path)) {
+    return `${url}?sanitize=true`;
+  }
+
+  return url;
 }
 
 export function docRef(path: string) {