+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Piral Changelog
 
-## 0.6.2 (to be determined)
+## 0.6.3 (to be determined)
+
+* Allow pilet scaffolding from non-NPM sources (#53)
+* Improved documentation code generation (#52)
+* Support configuration for `piral-ext` extensions (#51)
+* Remove output folder first on publish with `--fresh`
+
+## 0.6.2 (August 3, 2019)
 
 * Added `extendSharedDependencies` helper (#48)
 * Distinguish between Piral development and contributing (#42)